Update responsive UI Mobile

+++ database/migrations/2025_12_11_144243_create_submission_table.php
@@ -12,19 +12,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('submission', function (Blueprint $table) {
-            $table->id('id_submission');
-            $table->string('file_tugas');
-            $table->foreignId('id_tugas');
-            $table->foreignId('id_mahasiswa');
-            $table->enum('status', ['Dikumpulkan', 'Terlambat', 'Belum Mengumpulkan','Sudah Dinilai'])->default('Belum Mengumpulkan');
-            $table->integer('nilai')->nullable();
-            $table->text('catatan')->nullable();
-            $table->dateTime('submitted_at');
-            $table->foreign('id_tugas')->references('id_tugas')->on('tugas')->onDelete('cascade');
-            $table->foreign('id_mahasiswa')->references('id_mahasiswa')->on('mahasiswa')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('submission')) {
+            Schema::create('submission', function (Blueprint $table) {
+                $table->id('id_submission');
+                $table->string('file_tugas');
+                $table->foreignId('id_tugas');
+                $table->foreignId('id_mahasiswa');
+                $table->enum('status', ['Dikumpulkan', 'Terlambat', 'Belum Mengumpulkan','Sudah Dinilai'])->default('Belum Mengumpulkan');
+                $table->integer('nilai')->nullable();
+                $table->text('catatan')->nullable();
+                $table->dateTime('submitted_at');
+                $table->foreign('id_tugas')->references('id_tugas')->on('tugas')->onDelete('cascade');
+                $table->foreign('id_mahasiswa')->references('id_mahasiswa')->on('mahasiswa')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
